cdk_requirements.txt
aws-cdk.core
aws-cdk.aws-s3
aws-cdk.aws-ec2
aws-cdk.aws-iam
aws-cdk.aws-rds
aws-cdk.aws-ecs
